requirements.txt
annotated-types==0.7.0
bcrypt==4.3.0
cryptography==45.0.6
fastapi==0.116.1
httpcore==1.0.9
httptools==0.6.4
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.14.0
pyodbc==5.2.0
python-dotenv==1.1.1
python-jose==3.5.0
SQLAlchemy==2.0.43
sqlmodel==0.0.24
urllib3==2.5.0
uvicorn==0.35.0

